refactor(formula): extracted clean_formula
